Login with jwt (#28)
* FriendshipTest.java 업데이트 (#17)

* FriendshipTest.java 업데이트

* FriendshipTest.java 업데이트 (#18)

* feature: security filter chain, custom user details, UsernamePasswordAuthenticationFilter

+++ src/main/java/com/example/swiftgathering_server/service/MemberService.java
@@ -9,6 +9,7 @@ import com.example.swiftgathering_server.repository.MemberRepository;
 import com.example.swiftgathering_server.exception.AuthenticationException;
 import jakarta.persistence.EntityExistsException;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -18,6 +19,7 @@ import org.springframework.transaction.annotation.Transactional;
 public class MemberService {
 
     private final MemberRepository memberRepository;
+    private final BCryptPasswordEncoder bCryptPasswordEncoder;
 
     public Long register(RegisterDto registerDto) {
         memberRepository.findByLoginId(registerDto.getLoginId())
@@ -25,25 +27,30 @@ public class MemberService {
                     throw new EntityExistsException("Login ID already in use: " + registerDto.getLoginId());
                 });
 
+        String encodedPassword = bCryptPasswordEncoder.encode(registerDto.getLoginPassword());
         Member member = Member.builder()
                 .loginId(registerDto.getLoginId())
-                .loginPassword(registerDto.getLoginPassword())
+                .loginPassword(encodedPassword)
                 .name(registerDto.getName())
                 .build();
         return memberRepository.save(member);
     }
 
     public void resign(ResignDto resignDto) {
-        Member member = memberRepository.findByIdAndPassword(resignDto.getLoginId(), resignDto.getLoginPassword())
+        Member member = memberRepository.findByLoginId(resignDto.getLoginId())
                 .orElseThrow(() -> new AuthenticationException("Invalid login ID or password."));
+        if (!bCryptPasswordEncoder.matches(resignDto.getLoginPassword(), member.getLoginPassword())) {
+            throw new AuthenticationException("Invalid login ID or password.");
+        }
         memberRepository.remove(member);
     }
 
     public MyInfoDto verify(LoginDto loginDto) {
-        Long memberId = memberRepository
-                .findByIdAndPassword(loginDto.getLoginId(), loginDto.getLoginPassword())
-                .map(Member::getId)
+        Member member = memberRepository.findByLoginId(loginDto.getLoginId())
                 .orElseThrow(() -> new AuthenticationException("Invalid login ID or password."));
-        return new MyInfoDto(memberId);
+        if (!bCryptPasswordEncoder.matches(loginDto.getLoginPassword(), member.getLoginPassword())) {
+            throw new AuthenticationException("Invalid login ID or password.");
+        }
+        return new MyInfoDto(member.getId());
     }
 }
